+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Updated python dependencies.
+- Adds cEDH format to the list of formats.
+- Reorder the format list and customize the format titles better.
 
 ## [v11.4.1](https://github.com/lexicalunit/spellbot/releases/tag/v11.4.1) - 2024-10-03